+++ CHANGELOG.md
@@ -6,8 +6,9 @@ All notable changes to the code will be documented in this file.
 
 Features
 
-- added `Peacock: Darken` to the commands. Currently darkens the current color by 10%
-- added `Peacock: Ligthen` to the commands. Currently lightens the current color by 10%
+- added `Peacock: Darken` to the commands. Currently darkens the current color by the `darkenLightenPercentage` setting
+- added `Peacock: Ligthen` to the commands. Currently lightens the current color by the `darkenLightenPercentage` setting
+- added `peacock.darkenLightenPercentage` to settings. Indicates the percentage to darken or lighten the color (defaults to 10%)
 
 Breaking Changes